requirements.txt
playwright
google-generativeai
pillow